requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
qrcode[pil]
